requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-groq>=0.1.0
langgraph>=0.0.20
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
numpy>=1.24.0
PyGithub>=2.1.0
playwright>=1.40.0
duckduckgo-search>=3.9.0
httpx>=0.25.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.3
